refactor(console): move chart-pie to shared/components
