+++ gravitee-apim-console-webui/src/management/home/widgets/gio-api-lifecycle-state/gio-api-lifecycle-state.module.ts
@@ -18,7 +18,7 @@ import { CommonModule } from '@angular/common';
 
 import { GioApiLifecycleStateComponent } from './gio-api-lifecycle-state.component';
 
-import { GioChartPieModule } from '../../../../shared/components/widgets/gio-chart-pie/gio-chart-pie.module';
+import { GioChartPieModule } from '../../../../shared/components/gio-chart-pie/gio-chart-pie.module';
 
 @NgModule({
   imports: [CommonModule, GioChartPieModule],
