+++ gravitee-apim-console-webui/src/management/home/widgets/gio-api-state/gio-api-state.module.ts
@@ -18,7 +18,7 @@ import { CommonModule } from '@angular/common';
 
 import { GioApiStateComponent } from './gio-api-state.component';
 
-import { GioChartPieModule } from '../../../../shared/components/widgets/gio-chart-pie/gio-chart-pie.module';
+import { GioChartPieModule } from '../../../../shared/components/gio-chart-pie/gio-chart-pie.module';
 
 @NgModule({
   imports: [CommonModule, GioChartPieModule],
