+++ gravitee-apim-console-webui/src/shared/components/gio-chart-pie/gio-chart-pie.component.spec.ts
